Adding space for sealant in junctions, remove 45 option, top braces level
Cut glass with 45 degrees is prone to chip and hard to align.
This option existed in the code but not in the interface.

Adding support for configure the space for silicone (1.5mm default)

Mounting top brace level with top level, to make it easier to insert
silicone and eventually disassemble.